[orx-noise] Split NoiseFilters.kt, use emptyArray()
